Converted amount to double. Added more parameters for Regs. Forced position to portrait for all activities. Added: RecordsProvider; renamed recycler view.

+++ app/src/main/java/com/example/kotlinproject/fragments/AddRegFragmentAmount.kt
@@ -41,7 +41,7 @@ class AddRegFragmentAmount : Fragment() {
 
         nextButton.setOnClickListener{//@todo check
             findNavController().navigate(R.id.action_addRegFragmentAmount_to_addRegFragmentDetail)
-            (activity as AddRegActivity).amount = amountEditText.text.toString()
+            (activity as AddRegActivity).amount = amountEditText.text.toString().toDouble()
             Log.d("LifeCycle", "Presionado nextbutton")
         }
